# Tokenizer::scan performance improvements