@bobthecow 2019-07-21
I made the second change in mustache.php a while back and it made a huge difference: https://github.com/bobthecow/mustache.php/commit/6215e6b1b12477dd3686444ce081dbdfa20e3748

From my testing, it looks like the "tag change" logic does best when completely inlined, and it's not any harder to follow, so that's what I went with: https://github.com/bobthecow/mustache.php/commit/72d07520ec171c403f2d6d26a3ca33fa95be2c52